% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ecg.R
\name{process_ecg}
\alias{process_ecg}
\title{Identifies R-Spikes (via file annotation) in submitted ECG}
\usage{
process_ecg(file.list = c(), pattern = ".phys.gz",
  processing.mode = "batch", path = file.path(), recursive = TRUE,
  overwrite = FALSE, save.fecg = FALSE, check.ecg = FALSE,
  filter.width = 0.1, fl = 5, fh = 20, search = 10, threshold = 7,
  safe = 300)
}
\arguments{
\item{safe}{}
}
\description{
Identifies R-Spikes (via file annotation) in submitted ECG
}
\examples{


}
